backend/django/requirements.txt
asgiref==3.9.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cloudinary==1.44.1
cryptography==46.0.2
defusedxml==0.7.1
dj-database-url==3.0.1
dj-rest-auth==5.0.1
Django==5.2.6
django-cloudinary-storage==0.3.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.10
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
resend==2.17.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
